cyr AND lat for 3.3.0, validation success

+++ validate.sh
@@ -1,9 +1,26 @@
 #!/bin/bash
 
-# adjust these variables
-PHPBB_TAG="release-3.3.0"       # CHANGE the tag to checkout for validation HERE
-PHPBB_VERSION="3.3"             # CHANGE the main version of phpBB HERE - 3.0, 3.1, etc.
-PHPBB_LANGUAGE="sr-cyr"         # CHANGE the language iso to validate HERE
+# configuration, adjust as needed
+PHPBB_TAG="release-3.3.0"
+PHPBB_VERSION="3.3"
+# end of configuration
+
+function usage() {
+    echo "Validate Serbian translation build for phpBB."
+    echo "Usage:"
+    echo "    $0 LANG"
+    echo "Arguments:"
+    echo "    LANG: sr-cyr or sr-lat"
+    echo ""
+}
+
+if [[ -z $1 ]]; then
+    echo "ERROR: Missing required argument LANG."
+    usage
+    exit 0
+fi
+
+PHPBB_LANGUAGE=$1
 
 # prepare environment
 rm -rf $PHPBB_VERSION
@@ -13,8 +30,8 @@ mkdir -p $PHPBB_VERSION/$PHPBB_LANGUAGE/styles/prosilver/theme
 mkdir -p $PHPBB_VERSION/en/language
 mkdir -p $PHPBB_VERSION/en/styles/prosilver/theme
 
-cp -r src/language/$PHPBB_LANGUAGE $PHPBB_VERSION/$PHPBB_LANGUAGE/language/$PHPBB_LANGUAGE
-cp -r src/styles/prosilver/theme/$PHPBB_LANGUAGE $PHPBB_VERSION/$PHPBB_LANGUAGE/styles/prosilver/theme/$PHPBB_LANGUAGE
+cp -r build/language/$PHPBB_LANGUAGE $PHPBB_VERSION/$PHPBB_LANGUAGE/language/$PHPBB_LANGUAGE
+cp -r build/styles/prosilver/theme/$PHPBB_LANGUAGE $PHPBB_VERSION/$PHPBB_LANGUAGE/styles/prosilver/theme/$PHPBB_LANGUAGE
 
 git --git-dir=phpbb/.git --work-tree=phpbb checkout $PHPBB_TAG
